app/src/main/java/com/book/dan/beatbox/view/SoundViewModel.java
package com.book.dan.beatbox.view;

import android.databinding.BaseObservable;

import com.book.dan.beatbox.model.BeatBox;
import com.book.dan.beatbox.model.Sound;

public class SoundViewModel extends BaseObservable {
    private Sound mSound;
    private BeatBox mBeatBox;

    public SoundViewModel(BeatBox beatBox){
        mBeatBox = beatBox;
    }

    public String getTitle(){
        return mSound.getName();
    }

    public Sound getSound() {
        return mSound;
    }

    public void setSound(Sound sound) {
        mSound = sound;
        notifyChange();
    }

    public void onButtonClicked() {
        mBeatBox.play(mSound);
    }

}
